move kubestash usage functions into utils Signed-off-by: Anisur Rahman <anisur@appscode.com>
